refactor(ui): use f64 time values for sliders
standardize seek/time handling on f64 for iced sliders and round to milliseconds when saving playback progress

- change UI messages and state (SeekTo/PlayerTimeUpdated, current_time/total_duration) from i64 to f64
- update database progress writes to accept f64 and persist a rounded i64 position
- simplify proxy caches to use Rc/RefCell in the single-threaded UI context
- apply clippy-driven cleanups and update release/status documentation

+++ src/ui/message.rs
@@ -5,7 +5,7 @@ pub enum Message {
     // Player controls
     PlayPause,
     Stop,
-    SeekTo(i64),
+    SeekTo(f64),
     VolumeChanged(i32),
     SpeedChanged(f32),
 
@@ -33,7 +33,7 @@ pub enum Message {
 
     // Player events
     PlayerStateChanged(PlayerState),
-    PlayerTimeUpdated(i64),
+    PlayerTimeUpdated(f64),
     PlayerMediaParsed,
 
     // Background tasks
